analytics/requirements.txt
numpy==1.19.5
pandas==1.2.1
seaborn
matplotlib==3.3.3
textblob
googletrans
datetime
dtime
pysentimiento==0.1.0rc1
scikit-learn
gspread==3.6.0
gspread_dataframe==3.2.0
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib
wordcloud==1.8.1
plotly==4.14.1
kaleido==0.0.3.post1
oauth2client
python-dotenv
boto3
torch
botocore
pybase64
pydrive